BVAL-15 merged patches for " @Min and @Max fail with very high values of long"

+++ bval-jsr303/src/main/java/org/apache/bval/constraints/MinValidatorForNumber.java
@@ -43,7 +43,7 @@ public class MinValidatorForNumber implements ConstraintValidator<Min, Number> {
         } else if (value instanceof BigInteger) {
             return ((BigInteger) value).compareTo(BigInteger.valueOf(minValue)) != -1;
         } else {
-            return value.doubleValue() >= minValue;
+            return value.longValue() >= minValue;
         }
 
     }
